+++ src/CodeBuilder.ts
@@ -439,7 +439,10 @@ export abstract class CodeBuilder {
             if (tool == 'auto' || tool == undefined || tool == null) {
                 let hascpp = builderOptions.sourceList
                     .some((path) => /\.(?:cpp|c\+\+|cc|cxx)$/i.test(path));
-                tool = hascpp ? 'g++' : 'gcc';
+                if (toolchain.name == 'LLVM_ARM')
+                    tool = hascpp ? 'clang++' : 'clang'; // llvm-clang
+                else
+                    tool = hascpp ? 'g++' : 'gcc'; // gcc
             }
             // setup tool name
             builderOptions.options.linker['$toolName'] = tool;
@@ -888,22 +891,25 @@ export class ARMCodeBuilder extends CodeBuilder {
         */
         if (toolchain.name == 'GCC') {
             if (ArmCpuUtils.isArmArchName(cpu_id))
-                fpu_suffix = '';
+                fpu_suffix = ''; // 使用 march 时，无需 fpu_suffix 标记
         } else if (toolchain.name == 'AC6') {
             if (cpu_id.startsWith('armv8.1-m.'))
                 fpu_suffix = ''; // AC6 的 armv8.1-m 的 FPU 已经通过 +<扩展名> 开启
+        } else if (toolchain.name == 'LLVM_ARM') {
+            if (ArmCpuUtils.isArmArchName(cpu_id))
+                fpu_suffix = ''; // 使用 arch 时，无需 fpu_suffix 标记
         }
 
         options.global['microcontroller-cpu']   = cpu_id + fpu_suffix;
         options.global['microcontroller-fpu']   = cpu_id + fpu_suffix;
         options.global['microcontroller-float'] = cpu_id + fpu_suffix;
 
-        // 遗留参数，后面可能删除
-        options.global['target'] = cpu_id + fpu_suffix;
-
-        // arch extensions
-        if (ArmCpuUtils.isArmArchName(cpu_id)) {
-            const opts = config.compileConfig.archExtensions.split(',');
+        // ISA extensions
+        options.global['$arch-extensions'] = '';
+        options.global['$clang-arch-extensions'] = '';
+        options.global['$armlink-arch-extensions'] = '';
+        if (ArmCpuUtils.getArchExtensions(cpu_id, toolchain.name).length > 0) {
+            const opts = (config.compileConfig.archExtensions || '').split(',');
             // for gcc
             if (isGccFamilyToolchain(toolchain.name)) {
                 options.global['$arch-extensions'] = opts.join('');
@@ -913,6 +919,9 @@ export class ARMCodeBuilder extends CodeBuilder {
                 options.global['$clang-arch-extensions']   = opts.join('');
                 // 对于 linker 和 asm, 我们要将 '+' 替换成 '.' 字符
                 options.global['$armlink-arch-extensions'] = opts.map(v => v.replace('+', '.')).join('');
+            }
+            else if (toolchain.name == 'LLVM_ARM') {
+                // nothing todo now
             }
         }
 
@@ -943,28 +952,15 @@ export class ARMCodeBuilder extends CodeBuilder {
                 }
                 break;
 
-            // arm gcc
-            case 'GCC':
-                {
-                    scatterFilePath.split(',')
-                        .filter(s => s.trim() != '')
-                        .forEach((sctPath) => {
-                            ldFileList.push(this.convLinkerScriptPathForCompiler(sctPath));
-                        });
-                }
-                break;
-
-            // iar
-            case 'IAR_ARM':
-                {
-                    scatterFilePath.split(',')
-                        .filter(s => s.trim() != '')
-                        .forEach((sctPath) => {
-                            ldFileList.push(this.convLinkerScriptPathForCompiler(sctPath));
-                        });
-                }
-                break;
+            // by default
             default:
+                {
+                    scatterFilePath.split(',')
+                        .filter(s => s.trim() != '')
+                        .forEach((sctPath) => {
+                            ldFileList.push(this.convLinkerScriptPathForCompiler(sctPath));
+                        });
+                }
                 break;
         }